Implemented fetching of up to 100 comments within the last 24 hours. Comments displayed in list on the StockDetailScreen.kt

+++ app/src/main/java/io/xavier/topwsb/data/remote/dto/wsb_comments/WsbCommentDto.kt
@@ -24,7 +24,7 @@ data class WsbCommentDto(
     val collapsed_reason_code: Any,
     val comment_type: Any,
     val controversiality: Int,
-    val created_utc: Int,
+    val created_utc: Long,
     val distinguished: String,
     val gilded: Int,
     val gildings: Gildings,
