+++ app/src/main/java/io/xavier/topwsb/domain/model/WsbComment.kt
@@ -2,7 +2,7 @@ package io.xavier.topwsb.domain.model
 
 data class WsbComment(
     val text: String,
-    val createdUtc: Int,
+    val createdUtc: Long,
     val author: String,
     val permalink: String
 )
